カスタムデコレーターの作成

+++ backend/src/items/items.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateItemDto } from './dto/create-item.dto';
 import { ItemRepository } from './item.repository';
 import { Item } from 'src/entities/item.entity';
@@ -10,7 +14,9 @@ export class ItemsService {
 
   async findById(id: string): Promise<Item> {
     const item = await this.itemRepository.findOne(id);
-    if (!item) throw new NotFoundException();
+    if (!item) {
+      throw new NotFoundException('商品が見つかりません');
+    }
     return item;
   }
 
@@ -23,12 +29,22 @@ export class ItemsService {
     return item;
   }
 
-  async updateStatus(id: string): Promise<Item> {
+  async updateStatus(id: string, user: User): Promise<Item> {
     const item = await this.findById(id);
+
+    if (item.userId === user.id) {
+      throw new BadRequestException('自身の商品を購入することはできません');
+    }
+
     return await this.itemRepository.updateStatus(item);
   }
 
-  async delete(id: string): Promise<void> {
+  async delete(id: string, user: User): Promise<void> {
+    const item = await this.findById(id);
+
+    if (item.userId !== user.id) {
+      throw new BadRequestException('他人の商品を削除することはできません');
+    }
     this.itemRepository.delete(id);
   }
 }
